# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ETijucas API
  description: |
    API para o aplicativo ETijucas - Seu Dia a Dia.
    
    Este arquivo é a fonte de verdade para os contratos da API.
    O SDK TypeScript é gerado automaticamente a partir desta especificação.
  version: 1.0.0
  contact:
    name: ETijucas Team
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.etijucas.com.br
    description: Production server
...
paths:
...
  /api/v1/reports/{id}:
    get:
      tags: [Reports]
      summary: Get report details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Report details (public or private view depending on auth/ownership)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/PublicReport'
                      - $ref: '#/components/schemas/Report'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/reports/{id}/status:
    patch:
      tags: [Reports]
      summary: Update report status (admin/moderator)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, version]
              properties:
                status:
                  $ref: '#/components/schemas/ReportStatus'
                note:
                  type: string
                  nullable: true
                version:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Report'
        '409':
          description: Version conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  code:
                    type: string
                    example: REPORT_STATUS_CONFLICT
                  message:
                    type: string
                  currentVersion:
                    type: string
                    format: date-time
                  data:
                    $ref: '#/components/schemas/Report'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Resource not found.
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string
  schemas:
...
    ReportStatus:
      type: string
      enum: [recebido, em_analise, resolvido, rejeitado]
      description: |
        - recebido: Report received, awaiting analysis
        - em_analise: Report being analyzed
        - resolvido: Report resolved
        - rejeitado: Report rejected
    ReportCategory:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        color:
          type: string
        tips:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
    ReportMedia:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
        thumbUrl:
          type: string
        webUrl:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        mimeType:
          type: string
        size:
          type: integer
    ReportHistoryItem:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ReportStatus'
        note:
          type: string
          nullable: true
        at:
          type: string
          format: date-time
        by:
          type: string
          nullable: true
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        protocol:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ReportStatus'
        statusLabel:
          type: string
        statusColor:
          type: string
        statusIcon:
          type: string
        category:
          $ref: '#/components/schemas/ReportCategory'
        categoryId:
          type: string
          format: uuid
        addressText:
          type: string
          nullable: true
        addressSource:
          type: string
          nullable: true
        locationQuality:
          type: string
          nullable: true
        locationQualityLabel:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        locationAccuracyM:
          type: integer
          nullable: true
        bairroId:
          type: string
          format: uuid
          nullable: true
        media:
          type: array
          items:
            $ref: '#/components/schemas/ReportMedia'
        history:
          type: array
          items:
            $ref: '#/components/schemas/ReportHistoryItem'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    PublicReport:
      allOf:
        - $ref: '#/components/schemas/Report'
...